+++ pmd-html/src/main/java/net/sourceforge/pmd/lang/html/cpd/HtmlCpdLexer.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.html.ast;
+package net.sourceforge.pmd.lang.html.cpd;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -15,6 +15,10 @@ import net.sourceforge.pmd.lang.ast.Parser.ParserTask;
 import net.sourceforge.pmd.lang.ast.SemanticErrorReporter;
 import net.sourceforge.pmd.lang.document.TextDocument;
 import net.sourceforge.pmd.lang.html.HtmlLanguageModule;
+import net.sourceforge.pmd.lang.html.ast.ASTHtmlDocument;
+import net.sourceforge.pmd.lang.html.ast.ASTHtmlTextNode;
+import net.sourceforge.pmd.lang.html.ast.HtmlNode;
+import net.sourceforge.pmd.lang.html.ast.HtmlParser;
 
 /**
  * <p>Note: This class has been called HtmlTokenizer in PMD 6</p>.